+++ src/elements/speaker/linkGrid.svelte
